fixes for some warnings that FPC complained about.
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing-7.2@4001 d681b350-854c-439c-a81d-c0ce20710203

+++ src/dbc/ZDbcOracleResultSet.pas
@@ -2012,7 +2012,7 @@ begin
       FChunkSize, Len, False, FConSettings)
   else
     OraWriteLob(FPlainDriver, Buffer, FContextHandle, FErrorHandle, FLobLocator,
-      FChunkSize, Len+1, False, FConSettings);
+      FChunkSize, Int64(Len)+1, False, FConSettings);
 end;
 
 end.
